fix a couple ssues with 64bit NSInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40099 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSTableView.m
@@ -5552,7 +5552,7 @@ This method is deprecated, use -columnIndexesInRect:. */
   // Build a string of texts, each representing a line from the tableView,
   // separated by "\n":
   NSString *descriptionsList = nil;
-  unsigned int index = [rows firstIndex];
+  NSInteger index = [rows firstIndex];
   while (index != NSNotFound)
   {
     NSString *draggedItemString = [self _objectStringForTableColumn:[cols objectAtIndex:0]
